Retuns a conversion factor for the amount calculation from urine concentration and volume

You can get a conversion factor for the multiplication: conc * vol * factor = amount in the given unit.

UnitUrine(conU = "ng/mL", volU = "mL", amtU = "mg", MW = 0)

Arguments

conU
concentration unit

volU
volume unit

amtU
amount unit

MW
molecular weight

Value

Factor

conversion factor for multiplication with the unit in name

Examples

UnitUrine()

#>    mg 
#> 1e-06 

UnitUrine("ng/mL", "mL", "mg")

#>    mg 
#> 1e-06 

UnitUrine("ug/L", "mL", "mg")

#>    mg 
#> 1e-06 

UnitUrine("ug/L", "L", "mg")

#>    mg 
#> 0.001 

UnitUrine("ng/mL", "mL", "g")

#>     g 
#> 1e-09 

UnitUrine("ng/mL", "mL", "mol", MW=500)

#>   mol 
#> 2e-12 

UnitUrine("ng/mL", "mL", "mmol", MW=500)

#>  mmol 
#> 2e-09 

UnitUrine("ng/mL", "mL", "umol", MW=500)

#>  umol 
#> 2e-06