Table output of gAUCs

Do multiple AUCs and returns a result table. See gNCA for more detail i.e. iAUC

tblAUC(Data, key = "Subject", colX = "Time", colY = "Y", iAUC = "",
       Ymax = "Emax", XofYmax = "TEmax", AUCname = "AUEClast", Outer = "NEAREST")

Arguments

Data

data table name

key

column names of Data to be shown in the output table

colX

column name for x axis

colY

column name for y axis

iAUC

a data.frame to calculate interval AUCs

Ymax

usually Cmax or Emax

XofYmax

usually Tmax or TEmax

AUCname

usually AUClast or AUEClast

Outer

indicates how to do the out of x range point

Details

Tabular output of AUC with many subjects. This calculates only Cmax(Emax), Tmax(TEmax), AUCs

Value

Basically same with gAUC

Author

Kyun-Seop Bae <k@acr.kr>

See also

Examples

tblAUC(Theoph, key="Subject", colX="Time", colY="conc")
#>    Subject  Emax TEmax Tfrst Tlast  AUEClast
#> 1        1 10.50  1.12     0 24.37 148.92305
#> 2        2  8.33  1.92     0 24.30  91.52680
#> 3        3  8.20  1.02     0 24.17  99.28650
#> 4        4  8.60  1.07     0 24.65 106.79630
#> 5        5 11.40  1.00     0 24.35 121.29440
#> 6        6  6.44  1.15     0 23.85  73.77555
#> 7        7  7.09  3.48     0 24.22  90.75340
#> 8        8  7.56  2.02     0 24.12  88.55995
#> 9        9  9.03  0.63     0 24.43  86.32615
#> 10      10 10.21  3.55     0 23.70 138.36810
#> 11      11  8.00  0.98     0 24.08  80.09360
#> 12      12  9.75  3.52     0 24.15 119.97750

iAUC = data.frame(Name=c("AUC[0-12h]","AUC[0-24h]"), Start=c(0,0), End=c(12,24))
tblAUC(Indometh, key="Subject", colX="time", colY="conc", iAUC=iAUC)
#>   Subject Emax TEmax Tfrst Tlast AUEClast AUC.0.12h. AUC.0.24h.
#> 1       1 1.50  0.25  0.25     8  1.55375    2.12875    2.72875
#> 2       2 2.03  0.25  0.25     8  2.67875    3.50625    4.46625
#> 3       3 2.72  0.25  0.25     8  2.59375    3.59375    4.55375
#> 4       4 1.85  0.25  0.25     8  2.24625    2.98875    3.82875
#> 5       5 2.05  0.25  0.25     8  1.69750    2.45000    3.17000
#> 6       6 2.31  0.25  0.25     8  2.58375    3.52125    4.60125